+++ R/shapviz-package.R
@@ -4,8 +4,10 @@
 # Suppress R CMD check note
 #' @importFrom xgboost xgb.train
 
-globalVariables(c("from", "i", "id", "label", "to", "x", "shap", "SHAP",
-                  "feature", "value", "color", "Var2", "Var3", "S", "ind", "values"))
+globalVariables(c(
+  "from", "i", "id", "label", "to", "x", "shap", "SHAP",
+  "feature", "value", "color", "Var1", "Var2", "Var3", "S", "ind", "values"
+))
 
 .onLoad <- function(libname, pkgname) {
   op <- options()