+++ NEWS.md
@@ -3,7 +3,7 @@
 ### New feature
 
 `sv_interaction()`: New `kind = "bar"` to show mean absolute SHAP interactions/main effects as barplots.
-Modify via `fill` and `bar_width` arguments.
+Modify via `fill` and `bar_width` arguments [#169](https://github.com/ModelOriented/shapviz/pull/169).
 
 ### User-visible changes